diet_ai/dieta-backend/api/requirements.txt
# FastAPI Core
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database (usaremos asyncpg para async, psycopg2 para sync como fallback)
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9  # Como respaldo

# Pydantic
pydantic==2.5.0
pydantic-settings==2.1.0

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # Versión específica para evitar el error __about__
python-multipart==0.0.6

# Redis
redis==5.0.1

# Utilities
python-dotenv==1.0.0
httpx==0.25.2

# Development only (no en producción)
alembic==1.12.1  # Para migraciones
